bump

+++ include/memmap.h
@@ -23,7 +23,7 @@ PhysAddr getMapping(VirtAddr virtual, uint8_t *pageLevel);
  * \return 1 if slot found, 0 else
  */
 int findEmptySlotPageIdx(uint8_t targetType, uint16_t *idx);
-int findEmptyRangePageIdx(uint8_t targetType, uint16_t *idx, uint16_t count);
+int findEmptyRangePageIdx(uint8_t targetType, uint16_t *idx, size_t count);
 
 int mapPage(uint16_t *idx, uint8_t pageType, PhysAddr addr, uint64_t flags);
 int unmapPage(VirtAddr virt, PhysAddr *phys);
